% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gt.R
\name{apply_scale.gt_group}
\alias{apply_scale.gt_group}
\title{scale gt_group object}
\usage{
\method{apply_scale}{gt_group}(x, fontsize, tbl_scale, tbl_stub_pct)
}
\description{
scale gt_group object
}
\keyword{internal}
