% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gt.R
\name{apply_scale.gt_tbl}
\alias{apply_scale.gt_tbl}
\title{scale gt_tbl object}
\usage{
\method{apply_scale}{gt_tbl}(x, fontsize, tbl_scale, tbl_stub_pct)
}
\description{
scale gt_tbl object
}
\keyword{internal}
