% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_version}
\alias{check_pkg_version}
\title{Check package versions for docorator object are the same as loaded}
\usage{
check_pkg_version(x)
}
\arguments{
\item{x}{docorator object}
}
\description{
Check package versions for docorator object are the same as loaded
}
\keyword{internal}
