% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_chunks_all}
\alias{create_chunks_all}
\title{Create code chunks}
\usage{
create_chunks_all(x, transform)
}
\arguments{
\item{x}{docorator object}

\item{transform}{optional latex transformation function to apply to a gt latex string}
}
\value{
printed code chunk(s) to be included as-is in the render engine
}
\description{
Create code chunks
}
\examples{
\dontrun{
docorator <- gt::exibble |>
  gt::gt() |>
  as_docorator(save_object = FALSE)
create_chunks_all(docorator, transform = NULL)
}
}
\keyword{internal}
