% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc_datetime}
\alias{doc_datetime}
\title{Date and time of program run}
\usage{
doc_datetime()
}
\value{
character string
}
\description{
Date and time of program run
}
\examples{
\donttest{
 doc_datetime()
}
}
