% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc_pagenum}
\alias{doc_pagenum}
\title{Automatic page numbering}
\usage{
doc_pagenum()
}
\value{
character string containing latex code
}
\description{
Automatic page numbering
}
\examples{
doc_pagenum()

}
