% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc_path}
\alias{doc_path}
\title{Path of program}
\usage{
doc_path(filename = NULL, path = NULL)
}
\arguments{
\item{filename}{Name of output file}

\item{path}{program path}
}
\value{
character string
}
\description{
Path of program
}
\examples{
\donttest{
 doc_path(filename = "my_tbl.pdf", path = NULL)
}
}
