% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{fancyfoot}
\alias{fancyfoot}
\title{Construct document footer}
\usage{
fancyfoot(...)
}
\arguments{
\item{...}{Series of objects of class \code{fancyrow}. Each entry represents a new
row in the document footer.}
}
\value{
Character string containing latex code for the \code{fancyfoot} entries as
part of the \code{fancyhdr} latex framework
}
\description{
Define document footer through a series of \code{fancyrow}s. Each row represents
a new line in the footer with options for positioning text at left, center,
and/or right positions.
}
\examples{
fancyfoot(
 fancyrow(left = "My first footnote", right = doc_datetime())
)

}
