% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{fancyhead}
\alias{fancyhead}
\title{Construct document header}
\usage{
fancyhead(...)
}
\arguments{
\item{...}{Series of objects of class \code{fancyrow}. Each entry represents a new
row in the document header.}
}
\value{
Character string containing latex code for the \code{fancyhead} entries as
part of the \code{fancyhdr} latex framework
}
\description{
Define document header through a series of \code{fancyrow}s. Each row represents
a new line in the header with options for positioning text at left, center,
and/or right positions.
}
\examples{
fancyhead(
 fancyrow(left = "Protocol: 12345", right = doc_pagenum()),
 fancyrow(center = "Demographic Summary")
)

}
