% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{fancyrow}
\alias{fancyrow}
\title{Construct document header row}
\usage{
fancyrow(left = NA, center = NA, right = NA)
}
\arguments{
\item{left}{Character string to be aligned to the left side of the row.}

\item{center}{Character string to be aligned to the center of the row.}

\item{right}{Character string to be aligned to the right side of the row.}
}
\value{
Object of class \code{fancyrow}
}
\description{
Define a single row in the document header/footer. Each row represents
a single line of text, with options for positioning text at left, center,
and/or right.
}
\examples{
fancyrow(left = "Left most text", right = "Right most text")

fancyrow(center = "Just text in the center")

fancyrow(left = "All", center = "Three", right = "Positions filled")
}
