% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{gg_to_gt}
\alias{gg_to_gt}
\title{Convert ggplot object to gt from docorator object}
\usage{
gg_to_gt(x)
}
\arguments{
\item{x}{docorator object}
}
\description{
Convert ggplot object to gt from docorator object
}
\keyword{internal}
