% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{gg_to_image}
\alias{gg_to_image}
\title{ggplot to image filepath
Taken from code here: https://github.com/rstudio/gt/blob/6312cf5be92fae41633b5df7d41aa948b850aaf8/R/image.R#L351C1-L392C2}
\usage{
gg_to_image(plot_object, fig_dim = c(5, 8), path = getwd())
}
\arguments{
\item{plot_object}{ggplot object}

\item{fig_dim}{vector containing figure height and width in inches. Defaults to c(5,8) - docorator defaults}

\item{path}{file path including the png location to save the output to}
}
\value{
filepath to snap of ggplot
}
\description{
ggplot to image filepath
Taken from code here: https://github.com/rstudio/gt/blob/6312cf5be92fae41633b5df7d41aa948b850aaf8/R/image.R#L351C1-L392C2
}
\keyword{internal}
