% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{hf_escape}
\alias{hf_escape}
\title{escape the latex characters, but keep the output of doc_pagenum unescaped}
\usage{
hf_escape(x)
}
\arguments{
\item{x}{text string to be escaped}
}
\description{
escape the latex characters, but keep the output of doc_pagenum unescaped
}
