% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{hf_extract}
\alias{hf_extract}
\title{extract header footer info from docorator object}
\usage{
hf_extract(x)
}
\arguments{
\item{x}{a docorator object}
}
\value{
list of header footer info
}
\description{
extract header footer info from docorator object
}
\keyword{internal}
