% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{hf_height}
\alias{hf_height}
\alias{hf_height.default}
\alias{hf_height.fancyhdr}
\title{Calculate desired header or footer height for the document}
\usage{
hf_height(x, fontsize)

\method{hf_height}{default}(x, fontsize)

\method{hf_height}{fancyhdr}(x, fontsize)
}
\arguments{
\item{x}{header or footer}

\item{fontsize}{Document font size (pt)}
}
\value{
Numeric value
}
\description{
Calculate desired header or footer height for the document
}
\examples{
header <- fancyhead(
fancyrow(left = "Protocol: 12345", right = doc_pagenum()),
fancyrow(center = "Demographic Summary"))

hf_height(header, 10)

}
\keyword{internal}
