% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{hf_strip}
\alias{hf_strip}
\title{strip header footer info from ggplot object}
\usage{
hf_strip(x)
}
\arguments{
\item{x}{a ggplot object}
}
\value{
list of header footer info
}
\description{
strip header footer info from ggplot object
}
\keyword{internal}
