% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{png_path}
\alias{png_path}
\title{Path of png file}
\usage{
png_path(path = NULL)
}
\arguments{
\item{path}{path to png}
}
\value{
object with png attribute
}
\description{
Path of png file
}
\examples{
\dontrun{
png_path <- png_path(path = "path_to_my_png.png")
}
}
