% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{png_to_gt}
\alias{png_to_gt}
\title{Convert png object to gt from docorator object}
\usage{
png_to_gt(x)
}
\arguments{
\item{x}{docorator object}
}
\description{
Convert png object to gt from docorator object
}
\keyword{internal}
