% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj.R
\name{prep_obj.character}
\alias{prep_obj.character}
\title{prep object that is a character string (presumably latex code)}
\usage{
\method{prep_obj}{character}(x, transform = NULL, ...)
}
\description{
prep object that is a character string (presumably latex code)
}
\keyword{internal}
