% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj.R
\name{prep_obj.PNG}
\alias{prep_obj.PNG}
\title{prep object that is a path to a PNG}
\usage{
\method{prep_obj}{PNG}(x, transform = NULL, ...)
}
\description{
prep object that is a path to a PNG
}
\keyword{internal}
