% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_docx.R
\name{prep_obj_docx}
\alias{prep_obj_docx}
\alias{prep_obj_docx.default}
\alias{prep_obj_docx.PNG}
\alias{prep_obj_docx.gt_tbl}
\alias{prep_obj_docx.gt_group}
\title{Prepare table, listing, figure object for inclusion in the template Rmd}
\usage{
prep_obj_docx(x, ...)

\method{prep_obj_docx}{default}(x, ...)

\method{prep_obj_docx}{PNG}(x, ...)

\method{prep_obj_docx}{gt_tbl}(x, ...)

\method{prep_obj_docx}{gt_group}(x, ...)
}
\arguments{
\item{x}{docorator object containing info about the table, listing or figure}
}
\value{
xml to be inserted into word document
}
\description{
Prepare table, listing, figure object for inclusion in the template Rmd
}
\examples{
docorator <- gt::exibble |>
gt::gt() |>
as_docorator(
display_name = "mytbl", footer = NULL,
save_object = FALSE)

prepared_obj <- prep_obj_docx(docorator)
}
\keyword{internal}
