% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_rtf.R
\name{prep_obj_rtf.character}
\alias{prep_obj_rtf.character}
\title{prep object that is a character string (presumably latex code)}
\usage{
\method{prep_obj_rtf}{character}(x, ...)
}
\description{
prep object that is a character string (presumably latex code)
}
\keyword{internal}
