% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_rtf.R
\name{prep_obj_rtf.PNG}
\alias{prep_obj_rtf.PNG}
\title{prep object that is a path to a PNG}
\usage{
\method{prep_obj_rtf}{PNG}(x, ...)
}
\description{
prep object that is a path to a PNG
}
\keyword{internal}
