% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_tex.R
\name{prep_obj_tex}
\alias{prep_obj_tex}
\title{Prepare table, listing, figure object for inclusion in the template Rmd}
\usage{
prep_obj_tex(x, transform = NULL, ...)
}
\arguments{
\item{x}{table, listing, or figure object}

\item{transform}{optional latex transformation function to apply to a gt latex string}

\item{...}{additional args}
}
\value{
object to be included as-is in knitted rmd
}
\description{
Prepare table, listing, figure object for inclusion in the template Rmd
}
\keyword{internal}
