% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_docx}
\alias{render_docx}
\title{Render to docx}
\usage{
render_docx(x, display_loc = NULL, version_check = TRUE)
}
\arguments{
\item{x}{\code{docorator} object}

\item{display_loc}{optional path to save the output docx to}

\item{version_check}{Boolean indicating whether to print a note if gt or ggplot versions dont match between the original docorator object and the one being used for rendering}
}
\value{
This function saves a docx to a specified location
}
\description{
Render to docx
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{gt::gtcars |>
  dplyr::slice_head(n = 10) |>
  dplyr::select(mfr, model, year, msrp) |>
  gt::gt(groupname_col = "mfr",
         row_group_as_column = TRUE) |>
  as_docorator(
   header = fancyhead(fancyrow("Header 1"), fancyrow("Header 2")),
   display_name = "mytbl") |>
 render_docx()
}\if{html}{\out{</div>}}
}

