% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pdf.R
\name{render_pdf}
\alias{render_pdf}
\title{Render to pdf}
\usage{
render_pdf(
  x,
  display_loc = NULL,
  engine = c("latex", "html"),
  version_check = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{docorator} object}

\item{display_loc}{optional path to save the output pdf to}

\item{engine}{character vector of rendering engines to use. Options are "latex" (default) and "html".}

\item{version_check}{Boolean indicating whether to print a note if gt or ggplot versions dont match between the original docorator object and the one being used for rendering}

\item{...}{Additional arguments passed to the engine-specific render function.

For \code{engine = "latex"}, see \code{\link[=render_pdf_latex]{render_pdf_latex()}} for supported arguments.
For \code{engine = "html"}, see \code{\link[=render_pdf_html]{render_pdf_html()}} for supported arguments.}
}
\value{
This function saves a pdf to a specified location
}
\description{
Render to pdf
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{gt::gtcars |>
  dplyr::slice_head(n = 10) |>
  dplyr::select(mfr, model, year, msrp) |>
  gt::gt(groupname_col = "mfr",
         row_group_as_column = TRUE) |>
  as_docorator(
   header = fancyhead(fancyrow("Header 1"), fancyrow("Header 2")),
   display_name = "mytbl") |>
 render_pdf()
}\if{html}{\out{</div>}}
}

