% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gt.R
\name{apply_to_gt_group}
\alias{apply_to_gt_group}
\title{apply a gt function to a gt_group}
\usage{
apply_to_gt_group(func, args, call = rlang::caller_env())
}
\arguments{
\item{func}{string with function name}

\item{args}{named list of function arguments with gt_tbl or gt_group as first arg}

\item{call}{caller env}
}
\description{
apply a gt function to a gt_group
}
\examples{
gt_tbl <- gt::exibble|> gt::gt()
gt_group <- gt::gt_group(gt_tbl, gt_tbl)

func <- gt::tab_options
arg_list_group <- list(data = gt_group, page.header.use_tbl_headings = c(TRUE))

apply_to_gt_group(func,arg_list_group)

}
