% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_chunk}
\alias{create_chunk}
\title{Create prep_obj_tex code chunks
Taken from https://gist.github.com/MichaelJW/b4a3dd999a47b137d12f42a8f7562b11}
\usage{
create_chunk(x, transform)
}
\arguments{
\item{x}{docorator object}

\item{transform}{optional latex transformation function to apply to a gt latex string}
}
\value{
printed code chunk to be included as-is in the render engine
}
\description{
Create prep_obj_tex code chunks
Taken from https://gist.github.com/MichaelJW/b4a3dd999a47b137d12f42a8f7562b11
}
\examples{
\dontrun{
docorator <- gt::exibble |>
  gt::gt() |>
  as_docorator(save_object = FALSE)
create_chunk(docorator, transform = NULL)
}
}
\keyword{internal}
