% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{hf_process}
\alias{hf_process}
\alias{hf_process.default}
\alias{hf_process.character}
\alias{hf_process.fancyhead}
\alias{hf_process.fancyfoot}
\title{Process headers/footers}
\usage{
hf_process(x, escape_latex = TRUE, engine = "pdf")

\method{hf_process}{default}(x, escape_latex = TRUE, engine = "pdf")

\method{hf_process}{character}(x, escape_latex = TRUE, engine = "pdf")

\method{hf_process}{fancyhead}(x, escape_latex = TRUE, engine = "pdf")

\method{hf_process}{fancyfoot}(x, escape_latex = TRUE, engine = "pdf")
}
\arguments{
\item{x}{header or footer}

\item{escape_latex}{Boolean to escape latex in header/footer}

\item{engine}{render engine to dictate how the headers and footers should be constructed}
}
\value{
character string containing headers and footers latex code
}
\description{
Process headers/footers
}
\examples{
header <- fancyhead(
fancyrow(left = "Protocol: 12345", right = doc_pagenum()),
fancyrow(center = "Demographic Summary"))

hf_process(header)

}
\keyword{internal}
