% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fancyhdr.R
\name{hf_process}
\alias{hf_process}
\alias{hf_process.default}
\alias{hf_process.character}
\alias{hf_process.fancyhead}
\alias{hf_process.fancyfoot}
\title{Process headers/footers}
\usage{
hf_process(x, escape_latex = TRUE)

\method{hf_process}{default}(x, escape_latex = TRUE)

\method{hf_process}{character}(x, escape_latex = TRUE)

\method{hf_process}{fancyhead}(x, escape_latex = TRUE)

\method{hf_process}{fancyfoot}(x, escape_latex = TRUE)
}
\description{
Process headers/footers
}
\keyword{internal}
