% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_html.R
\name{prep_obj_html}
\alias{prep_obj_html}
\alias{prep_obj_html.default}
\alias{prep_obj_html.character}
\alias{prep_obj_html.gt_tbl}
\alias{prep_obj_html.gg}
\alias{prep_obj_html.PNG}
\alias{prep_obj_html.list}
\title{Prepare table, listing, figure object for inclusion in HTML}
\usage{
prep_obj_html(x, ...)

\method{prep_obj_html}{default}(x, ...)

\method{prep_obj_html}{character}(x, ...)

\method{prep_obj_html}{gt_tbl}(x, ...)

\method{prep_obj_html}{gg}(x, ...)

\method{prep_obj_html}{PNG}(x, ...)

\method{prep_obj_html}{list}(x, ...)
}
\arguments{
\item{x}{docorator object containing info about the table, listing or figure}

\item{...}{additional args}
}
\value{
object to be included as html in render
}
\description{
Prepare table, listing, figure object for inclusion in HTML
}
\examples{
docorator <- gt::exibble |>
gt::gt() |>
as_docorator(
display_name = "mytbl", footer = NULL,
save_object = FALSE)

prepared_obj <- prep_obj_html(docorator)

}
\keyword{internal}
