% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_obj_rtf.R
\name{prep_obj_rtf}
\alias{prep_obj_rtf}
\title{Prepare table, listing, figure object for output to rtf}
\usage{
prep_obj_rtf(x, ...)
}
\arguments{
\item{x}{docorator object containing display information}

\item{...}{additional args}
}
\value{
object to be included as-is in knitted rmd
}
\description{
Prepare table, listing, figure object for output to rtf
}
\keyword{internal}
