% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_html}
\alias{render_html}
\title{Render to HTML}
\usage{
render_html(x, display_loc = NULL)
}
\arguments{
\item{x}{\code{docorator} object}

\item{display_loc}{optional path to save the output pdf to}
}
\value{
Invisibly returns \code{file}. Called for its side effect of writing
an HTML file to disk.
}
\description{
Render to HTML
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{gt::gtcars |>
  dplyr::slice_head(n = 10) |>
  dplyr::select(mfr, model, year, msrp) |>
  gt::gt() |>
  as_docorator(
     display_name = "output",
     header = fancyhead(
       fancyrow(left = "Study ABC-123", right = "Draft"),
       fancyrow(center = "Table 1: Vehicle Summary")
     ),
     footer = fancyfoot(
       fancyrow(left = "Source: gtcars"),
       fancyrow(left = "program.R", right = format(Sys.Date(), "\%d\%b\%Y"))
     )
  ) |> 
  render_html()
}\if{html}{\out{</div>}}
}

